Move mock_connection to conftest
